Check all quizzes in a directory

Check the formatting of all quizzes in a given directory.

check_quizzes(quiz_dir = "quizzes", write_report = TRUE, verbose = TRUE)

Arguments

quiz_dir
A path to a directory full of quizzes that should all be checked with [ottr::check_all_quizzes].

write_report
TRUE/FALSE save warning report to a CSV file?

verbose
print diagnostic messages

Value

A list checks performed on each quiz

Examples

## Make a temporary quiz directory
quiz_dir <- dirname(good_quiz_path())

## Now check the quizzes in that directory
all_quiz_results <- check_quizzes(quiz_dir = quiz_dir)

#> 
#>  Checking quiz: /tmp/RtmpEsbVt7/temp_libpath1222c397ac8/ottr/extdata/quiz_bad.md

#> Checking question: It would be a shame  ... in quiz: quiz_bad.md

#> Warning: Colon detected in question on lines: 7 in question starting with:It would be a shame  ... in quiz: quiz_bad.md

#> Checking question: What a shame if we p ... in quiz: quiz_bad.md

#> Checking question: Shame shame if we sa ... in quiz: quiz_bad.md

#> Checking question: What if we don't put ... in quiz: quiz_bad.md

#> Warning: No correct answers provided for What if we don't put ... in quiz: quiz_bad.md

#> Checking question: What if only give co ... in quiz: quiz_bad.md

#> Warning: No incorrect answer options provided for What if only give co ... in quiz: quiz_bad.md

#> 
#>  Checking quiz: /tmp/RtmpEsbVt7/temp_libpath1222c397ac8/ottr/extdata/quiz_good.md

#> Checking question: First question to as ... in quiz: quiz_good.md

#> Checking question: Example without choo ... in quiz: quiz_good.md

#> Checking question: Question example wit ... in quiz: quiz_good.md

#> Checking question: A more complicated e ... in quiz: quiz_good.md

#> 
#>  Question error report saved to 'question_error_report.tsv'